practice_of_information_systems_1-main/group7/pyproject.toml
[tool.poetry]
name = "group7-backend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.75.2"
uvicorn = {extras = ["standard"], version = "^0.17.6"}
SQLAlchemy = "1.4.20"
aiomysql = "^0.1.0"
torch = "^1.11.0"
torchvision = "^0.12.0"
Pillow = "^9.1.1"
opencv-python = "^4.5.5"
pandas = "^1.4.2"
tqdm = "^4.64.0"
protobuf = "^4.21.1"
matplotlib = "^3.5.2"
seaborn = "^0.11.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
